# public/api/openapi.yaml
openapi: 3.0.0
info:
  title: 'Vigeogamesite API'
  version: 0.3.3
servers:
  -
    url: 'http://vgsite/api'
paths:
...
  '/users/{id}':
    get:
      description: 'A user'
      operationId: 'Users:GetOne'
      parameters:
        -
          $ref: '#/components/parameters/id'
        -
          $ref: '#/components/parameters/fields'
      responses:
        '200':
          description: Success!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
        '404':
          description: 'Requested user not found'
    patch:
      description: 'Modify a user'
      operationId: 'Users:Patch'
      parameters:
        -
          $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user'
      responses:
        '200':
          description: 'User modified'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: 'Requested user not found'
        '409':
          description: 'Conflict: Parameter not valid'
  /users:
    get:
      description: 'A list of users'
      operationId: 'Users:GetAll'
      parameters:
        -
          $ref: '#/components/parameters/page'
        -
          $ref: '#/components/parameters/per_page'
        -
          $ref: '#/components/parameters/sort'
        -
          $ref: '#/components/parameters/fields'
        -
          $ref: '#/components/parameters/q'
      responses:
        '200':
          description: Success!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
  /users/:
    post:
      description: 'Create a user'
      operationId: 'Users:Create'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user'
      responses:
        '200':
          description: 'User modified'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '409':
          description: 'Conflict: Parameter not valid'
components:
  schemas:
...
    user:
      properties:
        id:
          type: integer
        username:
          type: string
        password:
          description: 'A hashed string; Only given if explicitly requested using the `fields` parameter.'
          type: string
        email:
          type: string
        verified:
          type: boolean
        gender:
          description: 'enum(''he'', ''she'', ''them'')'
          type: string
        region:
          description: 'enum(''us'', ''jp'', ''eu'', ''au'')'
          type: string
        rank:
          type: integer
        avatar:
          type: string
        timezone:
          type: string
        data_created:
          type: date-time
        data_modified:
          type: date-time
        activity:
          type: date-time
        previous_activity:
          type: date-time
        href:
          type: string
      type: object
  parameters:
    q:
      name: q
      in: query
      description: 'Search term'
      schema:
        type: string
    sort:
      name: sort
      in: query
      description: 'Custom sorted results. Format: `?sort=fieldname[:asc|desc]`'
      schema:
        type: string
      example: 'release:desc'
    fields:
      name: fields
      in: query
      description: 'A list of comma-separated fields to include in the response object.'
      schema:
        type: string
      example: 'title,release_date,tags'
    page:
      name: page
      in: query
      description: 'Page number'
      schema:
        type: integer
    per_page:
      name: per_page
      in: query
      description: 'Number of results per page'
      schema:
        type: integer
    id:
      name: id
      in: path
      description: 'The numeric ID for the requested item'
      required: true
      schema:
        type: integer